Refactor MxPtrList to use p_ownership
