+++ LEGO1/mxdsselectaction.cpp
@@ -5,6 +5,7 @@
 #include "mxvariabletable.h"
 
 DECOMP_SIZE_ASSERT(MxDSSelectAction, 0xb0)
+DECOMP_SIZE_ASSERT(MxListEntry<MxString>, 0x18)
 
 // FUNCTION: LEGO1 0x100cb2b0
 MxDSSelectAction::MxDSSelectAction()
