+++ LEGO1/mxmediamanager.cpp
@@ -7,6 +7,8 @@
 #include "mxticklemanager.h"
 
 DECOMP_SIZE_ASSERT(MxMediaManager, 0x2c);
+DECOMP_SIZE_ASSERT(MxPresenterList, 0x18);
+DECOMP_SIZE_ASSERT(MxPresenterListCursor, 0x10);
 
 // FUNCTION: LEGO1 0x100b84c0
 MxMediaManager::MxMediaManager()
