+++ LEGO1/mxpresenter.h
@@ -10,7 +10,6 @@
 
 class MxCompositePresenter;
 class MxStreamController;
-class MxPresenter;
 
 // VTABLE: LEGO1 0x100d4d38
 // SIZE 0x40
