--- LEGO1/mxstringlist.cpp
@@ -1,5 +1,0 @@
-#include "mxstringlist.h"
-
-#include "decomp.h"
-
-DECOMP_SIZE_ASSERT(MxListEntry<MxString>, 0x18)
